+++ container-engine-lib/lib/backend_interface/kurtosis_backend.go
@@ -6,6 +6,9 @@ import (
 	"github.com/kurtosis-tech/container-engine-lib/lib/backend_interface/objects/enclave"
 	"github.com/kurtosis-tech/container-engine-lib/lib/backend_interface/objects/engine"
 	"github.com/kurtosis-tech/container-engine-lib/lib/backend_interface/objects/exec_result"
+	"github.com/kurtosis-tech/container-engine-lib/lib/backend_interface/objects/files_artifact"
+	"github.com/kurtosis-tech/container-engine-lib/lib/backend_interface/objects/files_artifact_expander"
+	"github.com/kurtosis-tech/container-engine-lib/lib/backend_interface/objects/files_artifact_expansion_volume"
 	"github.com/kurtosis-tech/container-engine-lib/lib/backend_interface/objects/module"
 	"github.com/kurtosis-tech/container-engine-lib/lib/backend_interface/objects/networking_sidecar"
 	"github.com/kurtosis-tech/container-engine-lib/lib/backend_interface/objects/port_spec"
@@ -45,8 +48,8 @@ type KurtosisBackend interface {
 		filters *engine.EngineFilters,
 	) (
 		successfulEngineIds map[string]bool, // "set" of engine IDs that were successfully stopped
-		erroredEngineIds map[string]error, // "set" of engine IDs that errored when stopping, with the error
-		resultErr error, // Represents an error with the function itself, rather than the engines
+		erroredEngineIds map[string]error,   // "set" of engine IDs that errored when stopping, with the error
+		resultErr error,                     // Represents an error with the function itself, rather than the engines
 	)
 
 	// Destroys the engines matching the given filters, regardless of if they're running or not
@@ -55,8 +58,8 @@ type KurtosisBackend interface {
 		filters *engine.EngineFilters,
 	) (
 		successfulEngineIds map[string]bool, // "set" of engine IDs that were successfully destroyed
-		erroredEngineIds map[string]error, // "set" of engine IDs that errored when destroying, with the error
-		resultErr error, // Represents an error with the function itself, rather than the engines
+		erroredEngineIds map[string]error,   // "set" of engine IDs that errored when destroying, with the error
+		resultErr error,                     // Represents an error with the function itself, rather than the engines
 	)
 
 	// Creates an enclave with the given enclave ID
@@ -114,7 +117,7 @@ type KurtosisBackend interface {
 		ipAddr net.IP, // TODO REMOVE THIS ONCE WE FIX THE STATIC IP PROBLEM!!
 		grpcPortNum uint16,
 		grpcProxyPortNum uint16,
-		enclaveDataDirpathOnHostMachine string,	// TODO DELETE WHEN WE HAVE AN ENCLAVE DATA VOLUME!
+		enclaveDataDirpathOnHostMachine string, // TODO DELETE WHEN WE HAVE AN ENCLAVE DATA VOLUME!
 		envVars map[string]string,
 	) (
 		*api_container.APIContainer,
@@ -146,7 +149,7 @@ type KurtosisBackend interface {
 		ctx context.Context,
 		filters *api_container.APIContainerFilters,
 	) (
-		// Successful & errored API containers are keyed by their enclave ID
+	// Successful & errored API containers are keyed by their enclave ID
 		successfulApiContainerIds map[enclave.EnclaveID]bool,
 		erroredApiContainerIds map[enclave.EnclaveID]error,
 		resultErr error,
@@ -163,7 +166,7 @@ type KurtosisBackend interface {
 		grpcPortNum uint16,
 		enclaveDataDirpathOnHostMachine string,
 		envVars map[string]string,
-	)(
+	) (
 		newModule *module.Module,
 		resultErr error,
 	)
@@ -183,7 +186,7 @@ type KurtosisBackend interface {
 		ctx context.Context,
 		filters *module.ModuleFilters,
 		shouldFollowLogs bool,
-	)(
+	) (
 		successfulModuleLogs map[module.ModuleGUID]io.ReadCloser,
 		erroredModuleGuids map[module.ModuleGUID]error,
 		resultError error,
@@ -195,8 +198,8 @@ type KurtosisBackend interface {
 		filters *module.ModuleFilters,
 	) (
 		successfulModuleIds map[module.ModuleGUID]bool, // "set" of module IDs that were successfully stopped
-		erroredModuleIds map[module.ModuleGUID]error, // "set" of module IDs that errored when being stopped, with the error
-		resultErr error, // Represents an error with the function itself, rather than the modules
+		erroredModuleIds map[module.ModuleGUID]error,   // "set" of module IDs that errored when being stopped, with the error
+		resultErr error,                                // Represents an error with the function itself, rather than the modules
 	)
 
 	// Destroys the modules with the given filters, regardless of if they're running or not
@@ -205,8 +208,8 @@ type KurtosisBackend interface {
 		filters *module.ModuleFilters,
 	) (
 		successfulModuleIds map[module.ModuleGUID]bool, // "set" of module IDs that were successfully destroyed
-		erroredModuleIds map[module.ModuleGUID]error, // "set" of module IDs that errored when destroying, with the error
-		resultErr error, // Represents an error with the function itself, rather than the modules
+		erroredModuleIds map[module.ModuleGUID]error,   // "set" of module IDs that errored when destroying, with the error
+		resultErr error,                                // Represents an error with the function itself, rather than the modules
 	)
 
 	// Creates a user service inside an enclave with the given configuration
@@ -223,7 +226,7 @@ type KurtosisBackend interface {
 		envVars map[string]string,
 		enclaveDataDirpathOnHostMachine string,
 		filesArtifactMountDirpaths map[string]string,
-    )(
+	) (
 		newUserService *service.Service,
 		resultErr error,
 	)
@@ -232,7 +235,7 @@ type KurtosisBackend interface {
 	GetUserServices(
 		ctx context.Context,
 		filters *service.ServiceFilters,
-	)(
+	) (
 		successfulUserServices map[service.ServiceGUID]*service.Service,
 		resultError error,
 	)
@@ -243,18 +246,18 @@ type KurtosisBackend interface {
 		ctx context.Context,
 		filters *service.ServiceFilters,
 		shouldFollowLogs bool,
-	)(
+	) (
 		successfulUserServiceLogs map[service.ServiceGUID]io.ReadCloser,
 		erroredUserServiceGuids map[service.ServiceGUID]error,
 		resultError error,
 	)
 
 	// Executes a shell command inside an user service instance indenfified by its ID
-	RunUserServiceExecCommands (
+	RunUserServiceExecCommands(
 		ctx context.Context,
 		enclaveId enclave.EnclaveID,
 		userServiceCommands map[service.ServiceGUID][]string,
-	)(
+	) (
 		succesfulUserServiceExecResults map[service.ServiceGUID]*exec_result.ExecResult,
 		erroredUserServiceGuids map[service.ServiceGUID]error,
 		resultErr error,
@@ -273,7 +276,7 @@ type KurtosisBackend interface {
 		initialDelayMilliseconds uint32, //The number of milliseconds to wait until executing the first HTTP call
 		retries uint32, //Max number of HTTP call attempts that this will execute until giving up and returning an error
 		retriesDelayMilliseconds uint32, //Number of milliseconds to wait between retries
-	)(
+	) (
 		resultErr error,
 	)
 
@@ -282,7 +285,7 @@ type KurtosisBackend interface {
 		ctx context.Context,
 		enclaveId enclave.EnclaveID,
 		serviceGUID service.ServiceGUID,
-	)(
+	) (
 		resultConn net.Conn,
 		resultErr error,
 	)
@@ -291,20 +294,20 @@ type KurtosisBackend interface {
 	StopUserServices(
 		ctx context.Context,
 		filters *service.ServiceFilters,
-	)(
+	) (
 		successfulUserServiceGuids map[service.ServiceGUID]bool, // "set" of user service GUIDs that were successfully stopped
-		erroredUserServiceGuids map[service.ServiceGUID]error, // "set" of user service GUIDs that errored when stopping, with the error
-		resultErr error, // Represents an error with the function itself, rather than the user services
+		erroredUserServiceGuids map[service.ServiceGUID]error,   // "set" of user service GUIDs that errored when stopping, with the error
+		resultErr error,                                         // Represents an error with the function itself, rather than the user services
 	)
 
 	// Destroy user services using the given filters,
 	DestroyUserServices(
 		ctx context.Context,
 		filters *service.ServiceFilters,
-	)(
+	) (
 		successfulUserServiceGuids map[service.ServiceGUID]bool, // "set" of user service GUIDs that were successfully destroyed
-		erroredUserServiceGuids map[service.ServiceGUID]error, // "set" of user service GUIDs that errored when destroying, with the error
-		resultErr error, // Represents an error with the function itself, rather than the user services
+		erroredUserServiceGuids map[service.ServiceGUID]error,   // "set" of user service GUIDs that errored when destroying, with the error
+		resultErr error,                                         // Represents an error with the function itself, rather than the user services
 	)
 
 	//Create a user service's  networking sidecar inside enclave
@@ -313,7 +316,7 @@ type KurtosisBackend interface {
 		enclaveId enclave.EnclaveID,
 		serviceGuid service.ServiceGUID,
 		ipAddr net.IP, // TODO REMOVE THIS ONCE WE FIX THE STATIC IP PROBLEM!!
-	)(
+	) (
 		*networking_sidecar.NetworkingSidecar,
 		error,
 	)
@@ -322,7 +325,7 @@ type KurtosisBackend interface {
 	GetNetworkingSidecars(
 		ctx context.Context,
 		filters *networking_sidecar.NetworkingSidecarFilters,
-	)(
+	) (
 		map[service.ServiceGUID]*networking_sidecar.NetworkingSidecar,
 		error,
 	)
@@ -332,7 +335,7 @@ type KurtosisBackend interface {
 		ctx context.Context,
 		enclaveId enclave.EnclaveID,
 		networkingSidecarsCommands map[service.ServiceGUID][]string,
-	)(
+	) (
 		successfulNetworkingSidecarExecResults map[service.ServiceGUID]*exec_result.ExecResult,
 		erroredUserServiceGuids map[service.ServiceGUID]error,
 		resultErr error,
@@ -342,7 +345,7 @@ type KurtosisBackend interface {
 	StopNetworkingSidecars(
 		ctx context.Context,
 		filters *networking_sidecar.NetworkingSidecarFilters,
-	)(
+	) (
 		successfulUserServiceGuids map[service.ServiceGUID]bool,
 		erroredUserServiceGuids map[service.ServiceGUID]error,
 		resultErr error,
@@ -352,9 +355,55 @@ type KurtosisBackend interface {
 	DestroyNetworkingSidecars(
 		ctx context.Context,
 		filters *networking_sidecar.NetworkingSidecarFilters,
-	)(
+	) (
 		successfulUserServiceGuids map[service.ServiceGUID]bool,
 		erroredUserServiceGuids map[service.ServiceGUID]error,
 		resultErr error,
+	)
+
+	//Create a files artifact exansion volume for user service and file artifact id
+	CreateFilesArtifactExpansionVolume(
+		ctx context.Context,
+		enclaveId enclave.EnclaveID,
+		serviceGuid service.ServiceGUID,
+		filesArtifactId files_artifact.FilesArtifactID,
+	) (
+		*files_artifact_expansion_volume.FilesArtifactExpansionVolume,
+		error,
+	)
+
+	//Destroy files artifact expansion volumes using the given filters
+	DestroyFilesArtifactExpansionVolumes(
+		ctx context.Context,
+		filters *files_artifact_expansion_volume.FilesArtifactExpansionVolumeFilters,
+	) (
+		successfulFileArtifactExpansionVolumeNames map[files_artifact_expansion_volume.FilesArtifactExpansionVolumeName]bool,
+		erroredFileArtifactExpansionVolumeNames map[files_artifact_expansion_volume.FilesArtifactExpansionVolumeName]error,
+		resultErr error,
+	)
+
+	//Run a files artifact expander
+	RunFilesArtifactExpander(
+		ctx context.Context,
+		guid files_artifact_expander.FilesArtifactExpanderGUID,
+		enclaveId enclave.EnclaveID,
+		filesArtifactExpansionVolumeName files_artifact_expansion_volume.FilesArtifactExpansionVolumeName,
+		enclaveDataDirpathOnHostMachine string,
+		destVolMntDirpathOnExpander string,
+		filesArtifactFilepathRelativeToEnclaveDatadirRoot string,
+		ipAddr net.IP, // TODO REMOVE THIS ONCE WE FIX THE STATIC IP PROBLEM!!
+	)(
+		*files_artifact_expander.FilesArtifactExpander,
+		error,
+	)
+
+	//Destroy files artifact expanders using the given filters
+	DestroyFilesArtifactExpanders(
+		ctx context.Context,
+		filters *files_artifact_expander.FilesArtifactExpanderFilters,
+	) (
+		successfulFilesArtifactExpanderGuids map[files_artifact_expander.FilesArtifactExpanderGUID]bool,
+		erroredFilesArtifactExpanderGuids map[files_artifact_expander.FilesArtifactExpanderGUID]error,
+		resultError error,
 	)
 }